+++ src/main/java/br/com/softlutions/contrutora/modules/obra/repositories/ObraRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.softlutions.contrutora.modules.obra.entities.Obra;
 
-public interface ObraRepository extends JpaRepository<Obra, Long> {
+public interface ObraRepository extends JpaRepository<Obra, Integer> {
 }
